datachannel: Make SendDataParams reliability fields absl::optional<uint16_t>
It doesn't make sense to use negative values or 0 to disable the feature,
so we use an optional 16bit value, reflecting the specification.
Values bigger than 16bit max are clamped down.

+++ api/transport/data_channel_transport_interface.h
@@ -47,16 +47,14 @@ struct SendDataParams {
   // If set, the maximum number of times this message may be
   // retransmitted by the transport before it is dropped.
   // Setting this value to zero disables retransmission.
-  // Must be non-negative. |max_rtx_count| and |max_rtx_ms| may not be set
-  // simultaneously.
-  absl::optional<int> max_rtx_count;
+  // |max_rtx_count| and |max_rtx_ms| may not be set simultaneously.
+  absl::optional<uint16_t> max_rtx_count;
 
   // If set, the maximum number of milliseconds for which the transport
   // may retransmit this message before it is dropped.
   // Setting this value to zero disables retransmission.
-  // Must be non-negative. |max_rtx_count| and |max_rtx_ms| may not be set
-  // simultaneously.
-  absl::optional<int> max_rtx_ms;
+  // |max_rtx_count| and |max_rtx_ms| may not be set simultaneously.
+  absl::optional<uint16_t> max_rtx_ms;
 };
 
 // Sink for callbacks related to a data channel.
